'makeComponentLogger' doc is incomplete
makeComponentLogger is a poor function name (it should probably be "getOrMakeComponentLogger"). The name implies that a new logger is always made each time it is called. This means that anyone who'd like to write their own convenience layer on top of the Zowe logger would think that there is no way to get access to a logger with a given name.

Hopefully the proposed documentation change will clarify this.