Repairing_Broken_Links_in_Whats_New_section
The existing syntax caused the links to break.
The broken links were in:
Version 2.6.1 (February 2023)
Version 2.3.1 (September 2022)
Version 2.0.0 (April 2022)
The earlier (incorrect links) are kept as they are in the existing sentences.
I have copied the sentences (as they are) and merely added the missing syntax. It should solve the problem.
I will remove the broken links from the files after we test the links and confirm that the problem of the broken links is solved.